Move MockHttpRequest to outer scope
